Tests/LibWeb/Text/expected/indexeddb-abort-in-success-with-queued-requests.txt

putA.success
tx.abort() called
putB.error: AbortError
putC.error: AbortError
tx.abort
DONE
